+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change Log
 
+<!--
 All notable changes to the "vscode-updater" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+-->
 
-## [Unreleased]
+## [1.1.0] - 2026-08-13
 
-- Initial release
+### Added
+- Extension icon for marketplace listing
+
+### Changed
+- Documented `installPath` safety restrictions in README
+- Removed `beta` from channel options